+++ src/ispd/arquivo/xml/models/builders/LoadBuilder.java
@@ -3,12 +3,12 @@ package ispd.arquivo.xml.models.builders;
 import ispd.arquivo.xml.utils.WrappedDocument;
 import ispd.arquivo.xml.utils.WrappedElement;
 import ispd.motor.random.TwoStageUniform;
-import ispd.motor.carga.workload.CollectionWorkloadGenerator;
-import ispd.motor.carga.workload.PerNodeWorkloadGenerator;
-import ispd.motor.carga.workload.GlobalWorkloadGenerator;
-import ispd.motor.carga.workload.TraceFileWorkloadGenerator;
-import ispd.motor.carga.workload.WorkloadGenerator;
-import ispd.motor.carga.workload.WorkloadGeneratorType;
+import ispd.motor.carga.impl.CollectionWorkloadGenerator;
+import ispd.motor.carga.impl.PerNodeWorkloadGenerator;
+import ispd.motor.carga.impl.GlobalWorkloadGenerator;
+import ispd.motor.carga.impl.TraceFileWorkloadGenerator;
+import ispd.motor.carga.impl.WorkloadGenerator;
+import ispd.motor.carga.impl.WorkloadGeneratorType;
 
 import java.io.File;
 import java.util.Optional;
