+++ src/ispd/gui/LoadConfigurationDialog.java
@@ -2,12 +2,12 @@ package ispd.gui;
 
 import ispd.arquivo.xml.TraceXML;
 import ispd.gui.auxiliar.MultipleExtensionFileFilter;
-import ispd.motor.carga.workload.CollectionWorkloadGenerator;
-import ispd.motor.carga.workload.PerNodeWorkloadGenerator;
-import ispd.motor.carga.workload.GlobalWorkloadGenerator;
-import ispd.motor.carga.workload.TraceFileWorkloadGenerator;
-import ispd.motor.carga.workload.WorkloadGenerator;
-import ispd.motor.carga.workload.WorkloadGeneratorType;
+import ispd.motor.carga.impl.CollectionWorkloadGenerator;
+import ispd.motor.carga.impl.PerNodeWorkloadGenerator;
+import ispd.motor.carga.impl.GlobalWorkloadGenerator;
+import ispd.motor.carga.impl.TraceFileWorkloadGenerator;
+import ispd.motor.carga.impl.WorkloadGenerator;
+import ispd.motor.carga.impl.WorkloadGeneratorType;
 
 import javax.swing.BorderFactory;
 import javax.swing.DefaultComboBoxModel;
